+++ python/Recorder.py
@@ -14,9 +14,9 @@ class Recorder:
         self.cfg = cfg
 
         # Get an appropriate file name and path.
-        now = datetime.now()
         self.video_path = self.cfg["video_storage"]
         self.ffmpeg_path = self.cfg["ffmpeg_path"]
+        self.hwe = self.cfg['hwe']
         self.file_name = f"{self.video_path}/{self.bracket}/{self.zoneID}.mp4"
 
     def get_path(self):
@@ -29,12 +29,17 @@ class Recorder:
 
     def start_recording(self):
         """Call ffmpeg to start the recording."""
-        self.start_time_seconds = round(time.time())
+        self.start_time_seconds = round(time.time())    
 
-        cmd = f'{self.cfg["ffmpeg_path"]} -y -thread_queue_size 1024         \
-        -f gdigrab -framerate 50 -video_size 1920x1080 -i desktop            \
-        -r 50 -preset fast -qp 23 -pix_fmt yuv420p                           \
-        "{self.file_name}"'
+        cmd = f'{self.cfg["ffmpeg_path"]} -y -thread_queue_size 1024 -f gdigrab -framerate 30 -video_size 1920x1080 -i desktop -r 30 -preset fast '
+
+        # Enable hardware encoding if possible. 
+        if (self.hwe == "NVIDIA"):
+            cmd = cmd + '-c:v h264_nvenc '
+        elif (self.hwe == "AMD"): 
+            cmd = cmd + '-c:v h264_amf '
+
+        cmd = cmd + f'-qp 23 -pix_fmt yuv420p "{self.file_name}"'
 
         # Mic capture -- requires setup
         #  -f dshow -i audio="Microphone (3- G533 Gaming Headset)"
@@ -42,12 +47,12 @@ class Recorder:
         # Audio capture -- requires setup
         #  -f dshow -i audio="virtual-audio-capturer"      
 
-        # Adds hardware encoding -- requires recent nvidia drivers (requires 11.1 or 471.41 not sure which is relevant)        
-        # -c:v h264_nvenc 
+        with open(f"{self.video_path}/diags/ffmpeg.log", "a") as ffmpeg_log:
 
-        with open(f"{self.video_path}/diags/ffmpeg.log", "a") as python_log:
-          self.recording_process = subprocess.Popen(
-              cmd, stdin=subprocess.PIPE, stdout=python_log, stderr=python_log, shell=True
+            ffmpeg_log.write("\n\nffmpeg commmand: " + cmd + "\n\n")
+
+            self.recording_process = subprocess.Popen(
+                cmd, stdin=subprocess.PIPE, stdout=ffmpeg_log, stderr=ffmpeg_log, shell=True
           )
 
         print("STARTED RECORDING", flush=True)